lib/gstreamer: add common encoder base
Add a encoder base class for gstreamer to share
common code between msdk and vaapi elements.
